% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{chi_squared}
\alias{chi_squared}
\title{Perform chi-squared test of y vs. x}
\usage{
chi_squared(x, outcome, weights)
}
\arguments{
\item{x}{The variable to consider splitting along}

\item{outcome}{The outcome variable, represented as a `matrix` where each column is an indicator for one level of the factor}

\item{weights}{A `vector` of prior weights}
}
\value{
The chi-squared test of independence betwen the proposed split variable and the binned variable
}
\description{
Perform chi-squared test of y vs. x
}
