% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_factor.R
\name{chi_squared_factor}
\alias{chi_squared_factor}
\title{perform chi-squared test of y vs. x}
\usage{
chi_squared_factor(x, outcome, weights)
}
\arguments{
\item{x}{A factor variable to be cross-tabulated with outcome}

\item{outcome}{A matrix of binned observed categories}

\item{weights}{A vector of prior weights}
}
\value{
Log p-value of the chi-squared statistic from the contingency table of x vs outcome.
}
\description{
perform chi-squared test of y vs. x
}
