% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.repart.R
\name{chisq.repart}
\alias{chisq.repart}
\title{A chi-square test that can work with observation weights and offsets}
\usage{
chisq.repart(x, ...)
}
\description{
A chi-square test that can work with observation weights and offsets
}
