% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_split.R
\name{do_split}
\alias{do_split}
\title{Return two new nodes, from splitting the data on a dichotomous variable}
\usage{
do_split(node, split)
}
\arguments{
\item{node}{A node of the tree}

\item{split}{A list that tells repart which variable to split the node on}
}
\value{
The node, changed from a leaf to a branch
}
\description{
Return two new nodes, from splitting the data on a dichotomous variable
}
