% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculate the classification entropy of a factor}
\usage{
entropy(x, outcome, weights)
}
\arguments{
\item{bins}{The outcome variable, represented as a data.frame where each column is an indicator for one level of the factor}
}
\value{
The information entropy of the binned variable
}
\description{
Calculate the classification entropy of a factor
}
