% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_factor.R
\name{entropy_factor}
\alias{entropy_factor}
\title{Calculate the classification entropy of a factor}
\usage{
entropy_factor(x, outcome, weights)
}
\arguments{
\item{x}{The factor variable to consider splitting along}

\item{outcome}{The outcome variable, represented as a `data.frame` where each column is an indicator for one level of the factor}

\item{weights}{A `vector` of prior weights}
}
\value{
The information entropy of the binned variable
}
\description{
Calculate the classification entropy of a factor
}
