% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_split.R
\name{find_split}
\alias{find_split}
\title{Identify which variable to split on, and where}
\usage{
find_split(
  response,
  data,
  outcome,
  splitfun,
  weights,
  adjust_threshold,
  alpha = 0.2
)
}
\arguments{
\item{response}{Name of the response variable}

\item{data}{The `data.frame` of data for this tree}

\item{outcome}{A `matrix` of possibly non-integer indicators for the outcome factor}

\item{splitfun}{The function that identifies the splits in the tree}

\item{weights}{A vector of prior weights}

\item{adjust_threshold}{A function that adjusts the outputs of `splitfun`}

\item{alpha}{Threshold for deciding whether to split}
}
\value{
`partysplit` for splitting at this node, or `NULL` if no split meets the criteria
}
\description{
Identify which variable to split on, and where
}
