% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{identity}
\alias{identity}
\title{Identity function just returns its parameter}
\usage{
identity(x)
}
\arguments{
\item{x}{Any variable}
}
\value{
The unmodified x
}
\description{
Identity function just returns its parameter
}
