% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{make_tree}
\alias{make_tree}
\title{Make a classification tree}
\usage{
make_tree(data, outcome)
}
\arguments{
\item{data}{The data frame containing all the input variables}

\item{outcome}{The known outcomes of the tree}
}
\value{
A classification tree to predict the distribution of the outcome based on the supplied data
}
\description{
Make a classification tree
}
