% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.repart.R
\name{residuals.repart}
\alias{residuals.repart}
\title{Calculate residuals for a classification tree}
\usage{
\method{residuals}{repart}(object, type = "logit", ...)
}
\arguments{
\item{type}{The typr of residuals to compute}

\item{...}{Other arguments passed to residuals}

\item{tree}{A classification tree fitted by repart}
}
\value{
A matrix of the classwise residuals
}
\description{
Calculate residuals for a classification tree
}
