% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_threshold.R
\name{sweep_threshold}
\alias{sweep_threshold}
\title{For splitting on a continuous variable, sweep the threshold across the possible values}
\usage{
sweep_threshold(x, outcome, weights, factor_function)
}
\arguments{
\item{x}{The factor variable to consider splitting along}

\item{outcome}{The outcome variable, represented as a `data.frame` where each column is an indicator for one level of the factor}

\item{weights}{A `vector` of prior weights}

\item{factor_function}{Function to call after `x` has been discretized at each threshold in the sweep}
}
\value{
A named list with elements `threshold`, which is the threshold that maximizes entropy gain and `gain`, which is the maximized entropy gain.
}
\description{
For splitting on a continuous variable, sweep the threshold across the possible values
}
