% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_split.R
\name{test_split}
\alias{test_split}
\title{Loop over the possible splits within this node. Returns a named list where the name is a variable name and the value is the node's entropy if a split is made on that variable.}
\usage{
test_split(node, minbin = 5)
}
\arguments{
\item{node}{The node that is being tested}

\item{minbin}{The minimum number of observations in a leaf node}
}
\value{
Maximum entropy that can be achieved by splitting at this node
}
\description{
Loop over the possible splits within this node. Returns a named list where the name is a variable name and the value is the node's entropy if a split is made on that variable.
}
