% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.repart.R
\name{fitted.repart}
\alias{fitted.repart}
\title{Extract the fitted values from the tree model}
\usage{
\method{fitted}{repart}(object, ...)
}
\arguments{
\item{...}{any additional parameters passed to `fitted`}

\item{tree}{The tree model from which to extract the fitted values}
}
\value{
The fitted values for the observations used to estimate the tree model
}
\description{
Extract the fitted values from the tree model
}
