% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{make_tree}
\alias{make_tree}
\title{Make a classification tree}
\usage{
make_tree(
  formula,
  data,
  outcome,
  splitfun,
  weights = NULL,
  adjust_threshold = identity,
  ...
)
}
\arguments{
\item{formula}{A `formula` describing the function to be estimated by the table}

\item{data}{A `data.frame` containing the data toi be used in fitting the model}

\item{outcome}{A `matrix` of binned outcomes for training the tree}

\item{splitfun}{The function to use in assessing each split}

\item{weights}{A `vector` of prior weights}

\item{adjust_threshold}{A function used to adjust the output of `splitfun` before assessing possible splits}
}
\value{
A classification tree to predict the distribution of the outcome based on the supplied data
}
\description{
Make a classification tree
}
