% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.repart.R
\name{residuals.repart}
\alias{residuals.repart}
\title{Calculate residuals for a classification tree}
\usage{
\method{residuals}{repart}(tree, type = "logit")
}
\arguments{
\item{tree}{A classification tree fitted by repart}

\item{type}{The typr of residuals to compute}
}
\value{
A matrix of the classwise residuals
}
\description{
Calculate residuals for a classification tree
}
