% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.repart.R
\name{table.repart}
\alias{table.repart}
\title{An analog of the built-in table function, but this one allows for incremental values in the cells to be summed up
(a behavior that's necessary for random-effect trees)}
\usage{
table.repart(y, x, weights)
}
\description{
An analog of the built-in table function, but this one allows for incremental values in the cells to be summed up
(a behavior that's necessary for random-effect trees)
}
